nodeSigner: move nodesigner.go to its own package
