+++ nodesigner/nodesigner.go
@@ -1,4 +1,4 @@
-package main
+package nodesigner
 
 import (
 	"fmt"
@@ -10,19 +10,19 @@ import (
 
 // nodeSigner is an implementation of the MessageSigner interface backed by the
 // identity private key of running lnd node.
-type nodeSigner struct {
+type NodeSigner struct {
 	privKey *btcec.PrivateKey
 }
 
 // newNodeSigner creates a new instance of the nodeSigner backed by the target
 // private key.
-func newNodeSigner(key *btcec.PrivateKey) *nodeSigner {
+func NewNodeSigner(key *btcec.PrivateKey) *NodeSigner {
 	priv := &btcec.PrivateKey{}
 	priv.Curve = btcec.S256()
 	priv.PublicKey.X = key.X
 	priv.PublicKey.Y = key.Y
 	priv.D = key.D
-	return &nodeSigner{
+	return &NodeSigner{
 		privKey: priv,
 	}
 }
@@ -30,7 +30,7 @@ func newNodeSigner(key *btcec.PrivateKey) *nodeSigner {
 // SignMessage signs a double-sha256 digest of the passed msg under the
 // resident node's private key. If the target public key is _not_ the node's
 // private key, then an error will be returned.
-func (n *nodeSigner) SignMessage(pubKey *btcec.PublicKey,
+func (n *NodeSigner) SignMessage(pubKey *btcec.PublicKey,
 	msg []byte) (*btcec.Signature, error) {
 
 	// If this isn't our identity public key, then we'll exit early with an
@@ -52,7 +52,7 @@ func (n *nodeSigner) SignMessage(pubKey *btcec.PublicKey,
 // SignCompact signs a double-sha256 digest of the msg parameter under the
 // resident node's private key. The returned signature is a pubkey-recoverable
 // signature.
-func (n *nodeSigner) SignCompact(msg []byte) ([]byte, error) {
+func (n *NodeSigner) SignCompact(msg []byte) ([]byte, error) {
 
 	// Otherwise, we'll sign the dsha256 of the target message.
 	digest := chainhash.DoubleHashB(msg)
@@ -72,4 +72,4 @@ func (n *nodeSigner) SignCompact(msg []byte) ([]byte, error) {
 
 // A compile time check to ensure that nodeSigner implements the MessageSigner
 // interface.
-var _ lnwallet.MessageSigner = (*nodeSigner)(nil)
+var _ lnwallet.MessageSigner = (*NodeSigner)(nil)
